Update problem 6 to add a static method
It doesn't make sense in my case to change any of these to static
methods since all of them use `self` to refer to the instance. However,
just for the sake of the exercise, I changed the `turn_engine_on` method
to a static method.

But it would still not be a good idea since an actual `turn_engine_on`
method would most likely need access to the attributes of a `Car`'s
instance.